requirements.txt
runpod>=1.6.0
llama-cpp-python>=0.2.0